requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dateutil==2.8.2
python-dotenv==1.2.1